feature: removePlaylist 관련 코드 수정 exception 패키지명 수정 api url 변경 관련 테스트 코드 수정

+++ src/main/java/com/team/comma/spotify/playlist/repository/PlaylistRepositoryCustom.java
@@ -8,7 +8,7 @@ public interface PlaylistRepositoryCustom {
 
     int findMaxListSequence();
 
-    int updateAlarmFlag(@Param("id") long id, @Param("alarmFlag") boolean alarmFlag);
+    long updateAlarmFlag(@Param("id") long id, @Param("alarmFlag") boolean alarmFlag);
 
-    int deletePlaylist(@Param("id") long id);
+    long deletePlaylist(@Param("id") long id);
 }
