+++ src/main/java/com/team/comma/spotify/playlist/repository/PlaylistRepositoryImpl.java
@@ -31,16 +31,16 @@ public class PlaylistRepositoryImpl implements PlaylistRepositoryCustom {
     }
 
     @Override
-    public int updateAlarmFlag(long id, boolean alarmFlag) {
-        return (int) queryFactory.update(playlist)
+    public long updateAlarmFlag(long id, boolean alarmFlag) {
+        return queryFactory.update(playlist)
                 .set(playlist.alarmFlag, alarmFlag)
                 .where(playlist.id.eq(id))
                 .execute();
     }
 
     @Override
-    public int deletePlaylist(long id) {
-        return (int) queryFactory.update(playlist)
+    public long deletePlaylist(long id) {
+        return queryFactory.update(playlist)
                 .set(playlist.delFlag,true)
                 .where(playlist.id.eq(id))
                 .execute();
